Added my own test level Added a new decorator to check the AI mech's battery life Created a new Behavior tree to test new AI behaviors

+++ Source/BatteryNAssault/BaseBattery.h
@@ -35,6 +35,6 @@ protected:
 
 public:
 	UPROPERTY(EditAnywhere)
-	int32 TeamID;
+	int8 TeamID;
 
 };
